Fix nozzle_throat_conditions build; add unit test
Grant Palmer reported that nozzle_throat_conditions would NaN on
Pleiades. Inspecting that program, I noticed it depends on lib/lewis
which is a mixed C/Fortran source library. Notes I left myself indicated
I had trouble transcibing the build of this library into CMake, because
if I linked to the global target (used to propagate common fortran
compiler flags), it would result in wacky flags for the *.c source files
included in the library. This probably indicates that I'm not using the
right CMake properties to propagate those options...

Regardless, the quick fix was to build the c-source in a separate lib
(clewis), then link the fortran source library to both clewis and
global. That worked, and the executable is producing good output with
GNU. A unit test was created to verify successful calculation.